refactor: extract neural-net module
Moves NeuralNetwork (Three.js / WebGL hero animation) and its
Canvas2D fallback NeuralNetwork2D out of js/main.js into the new
js/neural-net.js module.

Adds two supporting modules:
* js/three-context.js — single source of truth for the active THREE
  binding plus an `onChange` subscription so any module can stay in
  sync when tests swap THREE for a mock.
* js/utils.js — shared environment helpers (isLowPowerDevice,
  prefersReducedMotion, hasWebGLSupport, getTopoJSON) so that
  extracted modules don't need to reach back into main.js.

main.js now imports `NeuralNetwork` and `NeuralNetwork2D` from
neural-net.js and re-exports them; behaviour is unchanged.

207 tests still pass.  Vite build is unchanged.
